\name{plspm.formula-package}
\alias{plspm.formula-package}
\alias{plspm.formula.package}
\docType{package}
\title{Formula Based PLS Path Modeling}

\description{The main objective is to make easy the PLS Path Modeling with R using the package 'plspm'. It compute automatically the inner matrix and the outer list the 'plspm' function need simply by specify the model using formulas.
}

\details{
A formula based version of the package 'plspm'.The main fonction is 'plspm.formula' based on the \link[plspm]{plspm} function.
}

\author{
 ACHIEPO Odilon Yapo M. [aut, cre]
 ACHIEPO Odilon Yapo M. <kingodilon@gmail.com>
}

\references{
Gaston Sanchez, Laura Trinchera and Giorgio Russolillo (2015). plspm: Tools for Partial Least Squares Path Modeling (PLS-PM). R package version 0.4.7. \url{http://CRAN.R-project.org/package=plspm}
}

\keyword{package}

\seealso{
\code{\link[plspm.formula]{plspm.formula}}
}


