\name{plspmsat}
\alias{plspmsat}
\docType{data}
\title{The Satisfaction Dataset}

\description{
This data set contains the variables from a customer satisfaction study of a Spanish credit institution on 250 customers.
}

\usage{
data("plspmsat")
}

\details{
 See details on \link[plspm]{satisfaction}
}

\source{
Laboratory of Information Analysis and Modeling (LIAM). Facultat d'Informatica de Barcelona, Universitat Politecnica de Catalunya.
}

\keyword{datasets}
