\name{plspm.formula-package}
\alias{plspm.formula-package}
\alias{plspm.formula.package}
\docType{package}
\title{\packageTitle{plspm.formula}}

\description{
\packageDescription{plspm.formula}
}

\details{
A formula based version of the package 'plspm'.The main fonction is 'plspm.formula' based on the \link[plspm]{plspm} function.
}

\author{
\packageAuthor{plspm.formula} 
Maintainer: \packageMaintainer{plspm.formula}
}

\references{
Gaston Sanchez, Laura Trinchera and Giorgio Russolillo (2015). plspm: Tools for Partial Least Squares Path Modeling (PLS-PM). R package version 0.4.7. \url{http://CRAN.R-project.org/package=plspm}
}

\keyword{package}

\seealso{
\code{\link[plspm.formula]{plspm.formula}}
}


